refactor(Contexts): added application action handlers table
The application action handlers are functions that act on a given
application name. The action can be either global, as the open
application called from hs.application, or scoped to an application
instance, such as kill. This is necessary to prevent separate loops
or if checks on each loop iteraction to handle those scopes.